requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy==2.0.36
alembic==1.13.3

# Image processing
Pillow==10.4.0

# Document export
reportlab==4.2.5
python-docx==1.1.2
